Update README.md
At ALX School, we learn by doing. We have projects everyday that we have to push on GitHub in order for an automatic review to be done. This repository contains directories each corresponding to a system engineering or DevOps project. These projects are part of the ALX School year 1 curriculum. In these projects, we cover everything from the basics of the Linux Shell, Bash commands, Bash scripting.